apps/adk-agent-service/requirements.txt
google-adk
requests
python-dotenv
python-dotenv
